Move Capella defined predicates
