+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/logic/common/helpers/Predicates.java
@@ -13,13 +13,23 @@
 
 package tech.pegasys.teku.spec.logic.common.helpers;
 
+import org.apache.tuweni.bytes.Bytes;
 import org.apache.tuweni.bytes.Bytes32;
 import tech.pegasys.teku.infrastructure.crypto.Sha256;
 import tech.pegasys.teku.infrastructure.ssz.collections.SszBytes32Vector;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
+import tech.pegasys.teku.spec.config.SpecConfig;
 import tech.pegasys.teku.spec.datastructures.state.Validator;
 
 public class Predicates {
+
+  private final SpecConfig specConfig;
+
+  public static final Bytes ETH1_WITHDRAWAL_PREFIX = Bytes.fromHexString("0x01");
+
+  public Predicates(final SpecConfig specConfig) {
+    this.specConfig = specConfig;
+  }
 
   /**
    * Check if (this) validator is active in the given epoch.
@@ -63,5 +73,49 @@ public class Predicates {
     return !validator.isSlashed()
         && (validator.getActivationEpoch().compareTo(epoch) <= 0
             && epoch.compareTo(validator.getWithdrawableEpoch()) < 0);
+  }
+
+  /**
+   * Implementation of <b>has_eth1_withdrawal_credential</b> Capella Helper function. <br>
+   * Checks if validator has a 0x01 prefixed "eth1" withdrawal credential.
+   *
+   * @param validator the validator being checked
+   * @return true if the validator has an "eth1" withdrawal credential, false otherwise
+   */
+  public boolean hasEth1WithdrawalCredential(final Validator validator) {
+    return validator.getWithdrawalCredentials().slice(0, 1).equals(ETH1_WITHDRAWAL_PREFIX);
+  }
+
+  /**
+   * Implementation of <b>is_fully_withdrawable_validator</b> Capella helper function. <br>
+   * Checks if validator is fully withdrawable.
+   *
+   * @param validator the validator being checked
+   * @param balance the validator's balance
+   * @param epoch the current epoch
+   * @return true if the validator can fully withdraw their funds, false otherwise
+   */
+  public boolean isFullyWithdrawableValidator(
+      final Validator validator, final UInt64 balance, final UInt64 epoch) {
+    return hasEth1WithdrawalCredential(validator)
+        && validator.getWithdrawableEpoch().isLessThanOrEqualTo(epoch)
+        && balance.isGreaterThan(UInt64.ZERO);
+  }
+
+  /**
+   * Implementation of <b>is_partially_withdrawable_validator</b> Capella helper function. <br>
+   * Checks if validator is partially withdrawable.
+   *
+   * @param validator the validator being checked
+   * @param balance the validator's balance
+   * @return true if the validator can partially withdraw their funds, false otherwise
+   */
+  public boolean isPartiallyWithdrawableValidator(final Validator validator, final UInt64 balance) {
+    final UInt64 maxEffectiveBalance = specConfig.getMaxEffectiveBalance();
+    final boolean hasMaxEffectiveBalance =
+        validator.getEffectiveBalance().equals(maxEffectiveBalance);
+    final boolean hasExcessBalance = balance.isGreaterThan(maxEffectiveBalance);
+
+    return hasEth1WithdrawalCredential(validator) && hasMaxEffectiveBalance && hasExcessBalance;
   }
 }
